+++ src/value/flonum.rs
@@ -174,4 +174,4 @@ impl TryConvert for Flonum {
         }
     }
 }
-impl TryConvertOwned for Flonum {}
+unsafe impl TryConvertOwned for Flonum {}